Create statiscal_codes
These codes were developed in R software for statistical analysis and drawing up figures. The study is the analysis of recovery/regeneration of forest stretches that gallery the Cerrado, with different ages since the last disturbance (slash-and-burn agriculture).

We fitted multiple linear models to describe the SGF patches successional trajectory. We used fallow time, the adjusted eigenvalues of the PCA performed with the soil variables, and the water table depth class as predictive factors of the response variables. For each response variable, we selected the best-fit model using the linear, logarithmic, and polynomials (second-order) distributions. We selected the best model using the Akaike Information Criterion (AIC), and used the highest weight (wAICc) to select models with the highest probability of being the most parsimonious (Burnham et al., 2002). Subsequently, we verified the normality of residues, homogeneity of variances, and multicollinearity between the explanatory variables (Dormann et al., 2013; Draper & Smith, 1998). We used the Akaike weights (w) to determine the relative importance of each predictor variable selected from the global model (Barton, 2019). To assess differences in the floristic composition among the SGF patches of different ages, we performed an ordination analysis using the absolute abundance of individuals. We calculated the Bray-Curtis similarity index between pairs of sites, and input the resulting matrix in a non-metric multidimensional scaling analysis (NMDS, Quinn & Keough, 2002). We performed all analyses in program R (R Core Team, 2021).